Часть шестого домашнего задания.

+++ Assets/Scripts/Enemy/EnemyAsset.cs
@@ -6,7 +6,7 @@ namespace Enemy
     [CreateAssetMenu(menuName = "Assets/EnemyAsset", fileName = "EnemyAsset")]
     public class EnemyAsset : ScriptableObject
     {
-        public int StartHealth;
+        public float StartHealth;
         public bool IsFlyingEnemy;
         public float Speed;
